+++ user/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import redirect
 from django.urls import reverse
 from django.contrib import messages
+from django.contrib.auth.models import User
 from django.views.generic.edit import UpdateView
 from django.views import View
 from django.contrib.auth import authenticate, login, logout
@@ -15,6 +16,7 @@ from rest_framework.views import APIView
 from user.serializers import UserSerializer
 from user.models import StepProgress, CourseEnrollment
 from user.forms import EnrolForm, LoginForm
+
 
 
 def logout_view(request):
@@ -54,6 +56,7 @@ class Login(View):
     def post(self, request):
         form = self.form_class(request.POST)
         next = request.GET.get('next', '/')
+        errormsg = ''
         if form.is_valid():
             user = authenticate(
                 username=form.cleaned_data['username'],
@@ -62,8 +65,10 @@ class Login(View):
             if user is not None:
                 login(request, user)
                 return redirect(next)
-        message = 'Login failed!'
-        return render(request, self.template_name, context={'form': form, 'message': message})
+            else:
+                errormsg = 'Username and/or password are incorrect.'
+            
+        return render(request, self.template_name, context={'form': form, 'message': errormsg})
 
 
 class UserProfileApi(APIView):
@@ -79,6 +84,27 @@ class UserProfileApi(APIView):
         serializer = UserSerializer(user)
         return Response(serializer.data)
 
-def user_progress(request):
-    return render(request, 'user/admin_progress.html', context={'courses': CourseEnrollment.objects.all()})
+def user_progress_summary(request):
+    return render(request, 'user/admin_progress_summary.html', context={'courses': CourseEnrollment.objects.all()})
+
+def user_progress(request, user_id):
+    progress = []
+    u = User.objects.get(pk=user_id)
+    for course in u.courseenrollment_set.all():
+        data = {
+            'course': course.course.title,
+            'lessons': []
+        }
+        lessonids = [l['lesson__id'] for l in course.course.course_plan.values('lesson__id')]
+        for lesson in u.lessonprogress_set.filter(lesson__in=lessonids):
+            stepids = [ l['step'] for l in lesson.lesson.lesson_plan.values('step')]
+            steps = StepProgress.objects.filter(user=u, step__in=stepids)
+            lessondata = {
+                'lesson': lesson,
+                'steps': steps
+            }
+            data['lessons'].append(lessondata)
+        progress.append(data)
+    title = f"User Progress: {u.username} ({u.first_name} {u.last_name})"
+    return render(request, 'user/admin_user_progress.html', context={'courses': progress, 'user': u, 'title': title})