+++ changelog.md
@@ -1,3 +1,11 @@
+# Version 0.1.1
+
+- Fixing auth: user not initialised correctly
+- Log in errors show
+- Move homepage content into page content
+- Next and previous in steps and lessons working correctly
+- User progress in the admin site
+
 # Version 0.1.0
 
 - Initial release